Move authenticator package to internal
